Post-merge build fixes
Adapt PR's hash table and aggregation work to master's evolved APIs after
the master merge:

- `getDataAt` now returns `std::string_view` (not `StringRef`) — convert
  field accesses (`.data`, `.size`) to method calls in the new prefetch
  path in `HashMethod.h`.
- `IColumn::batchSerializeValueIntoMemory` now takes a
  `SerializationSettings *` parameter; thread one through `HashMethodABSerialized`.
- `AggregationMethodABStringNoCache::insertKeyIntoColumns` and
  `AggregationMethodABSerialized::insertKeyIntoColumns` similarly need to
  accept a `SerializationSettings *` to match the rest of the family that
  master extended.
- Add a default `static constexpr bool has_pre_computed_hashes = false;`
  on `HashMethodBase` so derived classes that don't opt into the prefetch
  path still satisfy `if constexpr (Derived::has_pre_computed_hashes)` in
  `findKey`.
- Restore `TopNFilterParameters` declaration in `MergeTreeIndices.h`.
  The PR's `bec285616a2` defined it there, but `MergeTreeIndices.h` was
  taken from master during conflict resolution. Reading-side code in
  `ReadFromMergeTree.h` (`setTopNFilterParams`, `top_n_filter_params`) and
  `MergeTreeDataSelectExecutor.h` still references the type.
- Cast `StringRefHash()(StringRef)` results to `UInt32` to satisfy
  `-Werror,-Wshorten-64-to-32`.

Build still fails further along — `optimizeOrderByLimit.cpp`,
`HashJoinMethodsImpl.h`, and a few other places need additional adaptation
to master's API changes that this commit doesn't yet address.

+++ src/Common/ColumnsHashing.h
@@ -630,7 +630,7 @@ struct HashMethodABSerialized
         {
             if (row_sizes[i] <= std::numeric_limits<UInt32>::max())
             {
-                auto h = StringRefHash()({memories[i] - row_sizes[i], row_sizes[i]});
+                auto h = static_cast<UInt32>(StringRefHash()({memories[i] - row_sizes[i], row_sizes[i]}));
                 hashes.getData()[i] = h;
                 if (row_sizes[i] < 12)
                     serialized_keys[i].low |= static_cast<uint32_t>(h);
